Restyle Calendar and Home pages, fix decimal weight logging
Replace native <select>/<input type=date> with styled Headless UI
components across the Daily Log page, group repeated exercise sets
under one header, unify secondary button styling, and give the
Weight Log section a proper heading plus a recent weigh-ins card so
it no longer sits alone. Restyle the Home dashboard charts to match
the app's card-based design system with a validated color palette.

Also fixes weight logging: the weight input's pattern blocked
decimals, and the weightlogs.weight column was INTEGER, so values
like 78.4kg silently failed or errored. Migrated the column to
DECIMAL(6,2) to match the precision already used for exercise weight.

+++ backend/models/weightlog.js
@@ -14,7 +14,7 @@ const WeightLog = sequelizeDB.define('weightlog', {
         allowNull: false,
     },
     weight: {
-        type: DataTypes.INTEGER
+        type: DataTypes.DECIMAL(6, 2)
     },
     logdate: {
         type: DataTypes.DATE,
